// src/app/cyprus/types.ts
export interface CyprusFlight {
  flightHashData: string
  currency: string
  flightType: string
  flightSegmentList: {
    origin: string
    destination: string
    flightType: string | null
    flightList: {
      departureDate: string
      departureTime: string
      arrivalDate: string
      arrivalTime: string
      flightDetails: {
        flightNo: string
        airline: string
        flightSource: string
        departureDate: string
        departureTime: string
        arrivalDate: string
        arrivalTime: string
        origin: string
        destination: string
        diffPrice?: ServicePriceType
        voyageCode: null
        className: string
        classType: string
        seat: number
        totalPrice: ServicePriceType
        key: null
        totalTax: ServicePriceType
        totalServiceFee: ServicePriceType
        priceCode: string
        basePrice: ServicePriceType
        taxes: ServicePriceType
        discount: ServicePriceType
        buyFee: {
          code: null
          price: ServicePriceType
        }
        fee: {
          code: null
          price: ServicePriceType
        }
        passengerPrices: null
        taxInfos: null
        serviceCharges: null
      }[]
    }[]
  }[]
}

export interface CyprusTransfer {
  transferHashData: string
  currency: string
  departureDate: string
  returnDate: string
  transferSegmentList: {
    transferTitle: string
    transferCode: string
    transferFrom: string
    transferTo: string
    key: null
    totalPrice: ServicePriceType
    diffPrice?: ServicePriceType
    priceCode: string
    basePrice: ServicePriceType
    taxes: ServicePriceType
    discount: ServicePriceType
    buyFee: ServiceFeePriceType
    fee: ServiceFeePriceType
    passengerPrices: null
    taxInfos: null
    serviceCharges: null
  }[]
}

export interface CyprusTransferApiResponse {
  flights: CyprusFlight | null
  transfers: CyprusTransfer | null
  hotelInfo: CyprusHotel | null
  sessionToken: string
  searchToken: string
  productKey: string | null
}